@@ -15,35 +15,34 @@
     "@turf/bbox": "^6.5.0",
     "axios": "^0.27.2",
     "colorbrewer": "^1.5.3",
-    "core-js": "^3.27.1",
+    "core-js": "^3.29.0",
     "file-saver": "^2.0.5",
-    "jose": "^4.11.2",
+    "jose": "^4.13.1",
     "mapbox-gl": "^1.13.2",
     "moment": "^2.29.4",
-    "papaparse": "^5.3.2",
-    "vue": "^2.7.14",
+    "papaparse": "^5.4.0",
+    "v-mapbox": "^3.3.1",
+    "vue": "^3.2.47",
     "vue-carbonbadge": "^1.2.1",
-    "vue-mapbox": "^0.4.1",
-    "vue-markdown-plus": "^2.1.1",
-    "vue-router": "^3.6.5",
-    "vuetify": "^2.6.13",
-    "vuex": "^3.6.2"
+    "vue-markdown-render": "^2.0.1",
+    "vue-router": "^4.1.6",
+    "vuetify": "^3.1.8",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.12",
+    "@babel/core": "^7.21.0",
     "@babel/eslint-parser": "^7.19.1",
     "@vue/cli-plugin-babel": "~5.0.0",
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "~5.0.8",
+    "@vue/compiler-sfc": "^3.2.47",
     "eslint": "^7.32.0",
     "eslint-plugin-vue": "^8.7.1",
     "gh-pages": "^3.2.3",
     "raw-loader": "^4.0.2",
     "sass": "~1.32.0",
     "sass-loader": "^10.4.1",
-    "vue-cli-plugin-vuetify": "~2.4.7",
-    "vue-template-compiler": "^2.7.14",
-    "vuetify-loader": "^1.9.2"
+    "vue-cli-plugin-vuetify": "~2.5.8"
   },
   "eslintConfig": {
     "root": true,
