@@ -1,17 +1,19 @@
-import Vue from "vue";
-import App from "./App.vue";
+import { createApp } from 'vue'
+import { createVuetify } from 'vuetify'
+import 'vuetify/styles'
+
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
 import router from "./router";
-import vuetify from "./plugins/vuetify";
+
 import store from "./plugins/store";
-import CarbonBadge from 'vue-carbonbadge'
-Vue.use(CarbonBadge)
 
-Vue.config.productionTip = false;
+import App from './App.vue'
 
-new Vue({
-  store,
-  beforeCreate() { this.$store.commit('initialiseStore') },
-  router,
-  vuetify,
-  render: h => h(App)
-}).$mount("#app");
+const vuetify = createVuetify({
+  components,
+  directives,
+})
+
+createApp(App).use(vuetify).use(router).use(store).mount('#app')
