@@ -6,7 +6,7 @@ export async function login(email) {
   try {
     const response = await axios.post(config.authenticate_path, { email: email, website: config.authenticate_website });
     if (response && response.status === 200) return true;
-  } catch (err) { 
+  } catch (err) {
     return false
   }
   return false;
