@@ -1,26 +1,23 @@
-import Vue from "vue";
-import Router from "vue-router";
+import { createRouter, createWebHashHistory } from 'vue-router'
 
-import Accessibility from "./views/accessibility-page.vue";
 import About from "./views/about-page.vue";
-import Home from "./views/home-page.vue";
+import Accessibility from "./views/accessibility-page.vue";
 import EventsData from "./views/events-data.vue";
+import Home from "./views/home-page.vue";
 import LibrariesData from "./views/libraries-data.vue";
 import LoansData from "./views/loans-data.vue";
+import Login from "./views/site-login.vue";
 import MembershipData from "./views/membership-data.vue";
 import MembershipMap from "./views/membership-map.vue";
 import MembershipPostcodeToLsoa from "./views/membership-postcode-to-lsoa.vue";
 import MobileLibraryStopsData from "./views/mobile-library-stops-data.vue";
 import PhysicalVisitsData from "./views/physical-visits-data.vue";
-import StockSummaryData from "./views/stock-summary-data.vue";
 import Privacy from "./views/privacy-page.vue";
-import Login from "./views/site-login.vue";
 import Profile from "./views/site-profile.vue";
+import StockSummaryData from "./views/stock-summary-data.vue";
 
-Vue.use(Router);
-
-export default new Router({
-  mode: "history",
+const router = createRouter({
+  history: createWebHashHistory(),
   base: process.env.BASE_URL,
   routes: [
     {
@@ -100,3 +97,5 @@ export default new Router({
     }
   ]
 });
+
+export default router;
