@@ -3,7 +3,7 @@
     <layout-header title="About" subtitle="What this site is for and how it works" />
     <section>
       <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
+        <vue-markdown :source="mdText"></vue-markdown>
       </v-container>
     </section>
   </div>
@@ -11,7 +11,7 @@
 <script>
 import Header from "../components/layout-header";
 import MarkDownData from "../markdown/about.md";
-import VueMarkdownPlus from "vue-markdown-plus";
+import VueMarkdown from "vue-markdown-render";
 
 export default {
   data() {
@@ -20,6 +20,6 @@ export default {
     };
   },
   methods: {},
-  components: { "layout-header": Header, VueMarkdownPlus }
+  components: { "layout-header": Header, VueMarkdown }
 };
 </script>
