@@ -6,7 +6,7 @@
     />
     <section>
       <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
+        <vue-markdown :source="mdText"></vue-markdown>
         <p class="text-center">
           <v-btn
             depressed
@@ -85,7 +85,7 @@
                 :loading="loadingServiceData"
                 loading-text="Loading... Please wait"
               >
-                <template v-slot:top>
+                <template #top>
                   <v-toolbar flat>
                     <v-spacer></v-spacer>
                     <v-dialog
@@ -96,7 +96,7 @@
                       content-class="elevation-0"
                     >
                       <template v-slot:activator="{ on, attrs }">
-                        <v-btn text color="success" v-bind="attrs" v-on="on">
+                        <v-btn v-bind="attrs" text color="success" v-on="on">
                           <v-icon left dark>mdi-plus-circle-outline</v-icon>
                           Add library
                         </v-btn>
@@ -104,207 +104,205 @@
                       <v-card>
                         <v-card-text>
                           <v-container>
-                            <template>
-                              <v-tabs>
-                                <v-tab>Library</v-tab>
-                                <v-tab>Location</v-tab>
-                                <v-tab>Opening hours</v-tab>
-                                <v-tab-item>
-                                  <v-container></v-container>
-                                  <v-row>
-                                    <v-col cols="12">
-                                      <v-text-field
-                                        outlined
-                                        v-model="editedItem.Library_name"
-                                        label="Name"
-                                      ></v-text-field>
-                                    </v-col>
-                                  </v-row>
-                                  <v-row>
-                                    <v-col cols="12" sm="8" md="8">
-                                      <v-select
-                                        v-model="editedItem.Type_of_library"
-                                        label="Type"
-                                        :items="[
-                                          { text: 'Local authority', value: 'LAL' },
-                                          {
-                                            text: 'Local authority (unstaffed)',
-                                            value: 'LAL-'
-                                          },
-                                          { text: 'Commisioned', value: 'CL' },
-                                          { text: 'Community-run', value: 'CRL' },
-                                          { text: 'Independent', value: 'IL' }
-                                        ]"
-                                        dense
-                                        outlined
-                                      ></v-select>
-                                    </v-col>
-                                    <v-col cols="12" sm="4" md="4">
-                                      <v-select
-                                        v-model="editedItem.Statutory"
-                                        :items="['Yes', 'No']"
-                                        label="Statutory"
-                                        dense
-                                        outlined
-                                      ></v-select>
-                                    </v-col>
-                                  </v-row>
-                                  <v-row>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Year_opened"
-                                        label="Year opened"
-                                        type="number"
-                                      ></v-text-field>
-                                    </v-col>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Year_closed"
-                                        label="Year closed"
-                                        type="number"
-                                      ></v-text-field>
-                                    </v-col>
-                                  </v-row>
-                                  <v-row>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Email_address"
-                                        label="Email address"
-                                      ></v-text-field>
-                                    </v-col>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.URL"
-                                        label="Website"
-                                      ></v-text-field>
-                                    </v-col>
-                                  </v-row>
-                                  <v-row>
-                                    <v-col cols="12" sm="12" md="12">
-                                      <v-textarea
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Notes"
-                                        label="Notes"
-                                        full-width
-                                      ></v-textarea>
-                                    </v-col>
-                                  </v-row>
-                                </v-tab-item>
-                                <v-tab-item>
-                                  <v-container></v-container>
-                                  <v-row>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Address_1"
-                                        label="Address 1"
-                                      ></v-text-field>
-                                    </v-col>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Address_2"
-                                        label="Address 2"
-                                      ></v-text-field>
-                                    </v-col>
-                                  </v-row>
-                                  <v-row>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Address_3"
-                                        label="Address 3"
-                                      ></v-text-field>
-                                    </v-col>
-                                    <v-col cols="12" sm="6" md="6">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="editedItem.Postcode"
-                                        label="Postcode"
-                                      ></v-text-field>
-                                    </v-col>
-                                  </v-row>
-                                  <v-row>
-                                    <v-col cols="12" sm="12" md="12">
-                                      <v-text-field
-                                        dense
-                                        outlined
-                                        v-model="
-                                          editedItem.Unique_property_reference_number
-                                        "
-                                        label="Unique property reference number"
-                                      ></v-text-field>
-                                    </v-col>
-                                  </v-row>
-                                </v-tab-item>
-                                <v-tab-item>
-                                  <v-container></v-container>
-                                  <v-row>
-                                    <v-col
-                                      cols="12"
-                                      sm="6"
-                                      md="6"
-                                      v-for="hourType in openingHourTypes"
-                                      :key="hourType"
-                                    >
-                                      <div v-for="day in days" :key="day">
-                                        <div class="text-subtitle-2">
-                                          <v-btn
-                                            icon
-                                            color="success"
-                                            v-on:click="
-                                              addOpeningHoursEntry(
-                                                day + '_' + hourType + '_hours'
-                                              )
-                                            "
-                                          >
-                                            <v-icon>mdi-plus-circle-outline</v-icon>
-                                          </v-btn>
-                                          {{ day }} {{ hourType }}
-                                        </div>
-
-                                        <v-chip
-                                          class="ma-1"
-                                          v-for="session in editedItem[
-                                            day + '_' + hourType + '_hours'
-                                          ]
-                                            ? editedItem[
-                                                day + '_' + hourType + '_hours'
-                                              ].split(',')
-                                            : ''"
-                                          :key="session"
-                                          small
-                                          label
-                                          close
-                                          close-icon="mdi-delete"
+                            <v-tabs>
+                              <v-tab>Library</v-tab>
+                              <v-tab>Location</v-tab>
+                              <v-tab>Opening hours</v-tab>
+                              <v-tab-item>
+                                <v-container></v-container>
+                                <v-row>
+                                  <v-col cols="12">
+                                    <v-text-field
+                                      outlined
+                                      v-model="editedItem.Library_name"
+                                      label="Name"
+                                    ></v-text-field>
+                                  </v-col>
+                                </v-row>
+                                <v-row>
+                                  <v-col cols="12" sm="8" md="8">
+                                    <v-select
+                                      v-model="editedItem.Type_of_library"
+                                      label="Type"
+                                      :items="[
+                                        { text: 'Local authority', value: 'LAL' },
+                                        {
+                                          text: 'Local authority (unstaffed)',
+                                          value: 'LAL-'
+                                        },
+                                        { text: 'Commisioned', value: 'CL' },
+                                        { text: 'Community-run', value: 'CRL' },
+                                        { text: 'Independent', value: 'ICL' }
+                                      ]"
+                                      dense
+                                      outlined
+                                    ></v-select>
+                                  </v-col>
+                                  <v-col cols="12" sm="4" md="4">
+                                    <v-select
+                                      v-model="editedItem.Statutory"
+                                      :items="['Yes', 'No']"
+                                      label="Statutory"
+                                      dense
+                                      outlined
+                                    ></v-select>
+                                  </v-col>
+                                </v-row>
+                                <v-row>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Year_opened"
+                                      label="Year opened"
+                                      type="number"
+                                    ></v-text-field>
+                                  </v-col>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Year_closed"
+                                      label="Year closed"
+                                      type="number"
+                                    ></v-text-field>
+                                  </v-col>
+                                </v-row>
+                                <v-row>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Email_address"
+                                      label="Email address"
+                                    ></v-text-field>
+                                  </v-col>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.URL"
+                                      label="Website"
+                                    ></v-text-field>
+                                  </v-col>
+                                </v-row>
+                                <v-row>
+                                  <v-col cols="12" sm="12" md="12">
+                                    <v-textarea
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Notes"
+                                      label="Notes"
+                                      full-width
+                                    ></v-textarea>
+                                  </v-col>
+                                </v-row>
+                              </v-tab-item>
+                              <v-tab-item>
+                                <v-container></v-container>
+                                <v-row>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Address_1"
+                                      label="Address 1"
+                                    ></v-text-field>
+                                  </v-col>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Address_2"
+                                      label="Address 2"
+                                    ></v-text-field>
+                                  </v-col>
+                                </v-row>
+                                <v-row>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Address_3"
+                                      label="Address 3"
+                                    ></v-text-field>
+                                  </v-col>
+                                  <v-col cols="12" sm="6" md="6">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="editedItem.Postcode"
+                                      label="Postcode"
+                                    ></v-text-field>
+                                  </v-col>
+                                </v-row>
+                                <v-row>
+                                  <v-col cols="12" sm="12" md="12">
+                                    <v-text-field
+                                      dense
+                                      outlined
+                                      v-model="
+                                        editedItem.Unique_property_reference_number
+                                      "
+                                      label="Unique property reference number"
+                                    ></v-text-field>
+                                  </v-col>
+                                </v-row>
+                              </v-tab-item>
+                              <v-tab-item>
+                                <v-container></v-container>
+                                <v-row>
+                                  <v-col
+                                    cols="12"
+                                    sm="6"
+                                    md="6"
+                                    v-for="hourType in openingHourTypes"
+                                    :key="hourType"
+                                  >
+                                    <div v-for="day in days" :key="day">
+                                      <div class="text-subtitle-2">
+                                        <v-btn
+                                          icon
                                           color="success"
-                                          v-on:click:close="
-                                            removeOpeningHoursEntry(
-                                              day + '_' + hourType + '_hours',
-                                              session
+                                          v-on:click="
+                                            addOpeningHoursEntry(
+                                              day + '_' + hourType + '_hours'
                                             )
                                           "
-                                          >{{ session }}</v-chip
                                         >
+                                          <v-icon>mdi-plus-circle-outline</v-icon>
+                                        </v-btn>
+                                        {{ day }} {{ hourType }}
                                       </div>
-                                    </v-col>
-                                  </v-row>
-                                </v-tab-item>
-                              </v-tabs>
-                            </template>
+
+                                      <v-chip
+                                        class="ma-1"
+                                        v-for="session in editedItem[
+                                          day + '_' + hourType + '_hours'
+                                        ]
+                                          ? editedItem[
+                                              day + '_' + hourType + '_hours'
+                                            ].split(',')
+                                          : ''"
+                                        :key="session"
+                                        small
+                                        label
+                                        close
+                                        close-icon="mdi-delete"
+                                        color="success"
+                                        v-on:click:close="
+                                          removeOpeningHoursEntry(
+                                            day + '_' + hourType + '_hours',
+                                            session
+                                          )
+                                        "
+                                        >{{ session }}</v-chip
+                                      >
+                                    </div>
+                                  </v-col>
+                                </v-row>
+                              </v-tab-item>
+                            </v-tabs>
                           </v-container>
                         </v-card-text>
                         <v-card-actions>
@@ -369,7 +367,7 @@
                                 v-model="openingHoursOpenMenuOpen"
                                 :close-on-content-click="false"
                                 :nudge-right="40"
-                                :return-value.sync="openingHoursOpen"
+                                :v-model:return-value="openingHoursOpen"
                                 transition="scale-transition"
                                 offset-y
                                 max-width="290px"
@@ -377,12 +375,12 @@
                               >
                                 <template v-slot:activator="{ on, attrs }">
                                   <v-text-field
+                                    v-bind="attrs"
                                     dense
                                     outlined
                                     v-model="openingHoursOpen"
                                     label="Open"
                                     readonly
-                                    v-bind="attrs"
                                     v-on="on"
                                   ></v-text-field>
                                 </template>
@@ -401,7 +399,7 @@
                                 v-model="openingHoursCloseMenuOpen"
                                 :close-on-content-click="false"
                                 :nudge-right="40"
-                                :return-value.sync="openingHoursClose"
+                                :v-model:return-value="openingHoursClose"
                                 transition="scale-transition"
                                 offset-y
                                 max-width="290px"
@@ -409,12 +407,12 @@
                               >
                                 <template v-slot:activator="{ on, attrs }">
                                   <v-text-field
+                                    v-bind="attrs"
                                     dense
                                     outlined
                                     v-model="openingHoursClose"
                                     label="Close"
                                     readonly
-                                    v-bind="attrs"
                                     v-on="on"
                                   ></v-text-field>
                                 </template>
@@ -447,7 +445,7 @@
                   </v-toolbar>
                 </template>
                 <template v-slot:[`item.Library_name`]="{ item }">
-                  <v-edit-dialog large :return-value.sync="item.Library_name">
+                  <v-edit-dialog large :v-model:return-value="item.Library_name">
                     {{ item.Library_name }}
                     <template v-slot:input>
                       <p>
@@ -462,7 +460,7 @@
                   </v-edit-dialog>
                 </template>
                 <template v-slot:[`item.Year_closed`]="{ item }">
-                  <v-edit-dialog large :return-value.sync="item.Year_closed">
+                  <v-edit-dialog large :v-model:return-value="item.Year_closed">
                     {{ item.Year_closed }}
                     <template v-slot:input>
                       <p>
@@ -478,7 +476,7 @@
                   </v-edit-dialog>
                 </template>
                 <template v-slot:[`item.Type_of_library`]="{ item }">
-                  <v-edit-dialog large :return-value.sync="item.Type_of_library">
+                  <v-edit-dialog large :v-model:return-value="item.Type_of_library">
                     {{ item.Type_of_library }}
                     <template v-slot:input>
                       <p>
@@ -504,7 +502,7 @@
                   </v-edit-dialog>
                 </template>
                 <template v-slot:[`item.Address_1`]="{ item }">
-                  <v-edit-dialog large :return-value.sync="item.Address_1">
+                  <v-edit-dialog large :v-model:return-value="item.Address_1">
                     {{ item.Address_1 }}
                     <template v-slot:input>
                       <p>
@@ -520,7 +518,7 @@
                   </v-edit-dialog>
                 </template>
                 <template v-slot:[`item.Postcode`]="{ item }">
-                  <v-edit-dialog large :return-value.sync="item.Postcode">
+                  <v-edit-dialog large :v-model:return-value="item.Postcode">
                     {{ item.Postcode }}
                     <template v-slot:input>
                       <p>
@@ -536,7 +534,7 @@
                   </v-edit-dialog>
                 </template>
                 <template v-slot:[`item.Statutory`]="{ item }">
-                  <v-edit-dialog large :return-value.sync="item.Statutory">
+                  <v-edit-dialog large :v-model:return-value="item.Statutory">
                     {{ item.Statutory }}
                     <template v-slot:input>
                       <p>
@@ -625,7 +623,7 @@ import * as Papa from "papaparse";
 import * as schemaHelper from "../helpers/schemas";
 
 import MarkDownData from "../markdown/librariesdata.md";
-import VueMarkdownPlus from "vue-markdown-plus";
+import VueMarkdown from "vue-markdown-render";
 
 export default {
   data() {
@@ -843,7 +841,7 @@ export default {
   components: {
     "layout-header": Header,
     "service-select": ServiceSelect,
-    VueMarkdownPlus
+    VueMarkdown
   }
 };
 </script>
