@@ -6,7 +6,7 @@
     />
     <section>
       <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
+        <vue-markdown :source="mdText"></vue-markdown>
         <p class="text-center">
           <v-btn
             depressed
@@ -25,7 +25,7 @@
 <script>
 import Header from "../components/layout-header";
 import MarkDownData from "../markdown/membershipdata.md";
-import VueMarkdownPlus from "vue-markdown-plus";
+import VueMarkdown from "vue-markdown-render";
 
 export default {
   data() {
@@ -36,7 +36,7 @@ export default {
   methods: {},
   components: {
     "layout-header": Header,
-    VueMarkdownPlus
+    VueMarkdown
   }
 };
 </script>
