@@ -3,7 +3,7 @@
     <layout-header title="Profile" subtitle="See access and profile information" />
     <section>
       <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
+        <vue-markdown :source="mdText"></vue-markdown>
       </v-container>
     </section>
     <section>
@@ -65,7 +65,7 @@
 <script>
 import Header from "../components/layout-header";
 import MarkDownData from "../markdown/profile.md";
-import VueMarkdownPlus from "vue-markdown-plus";
+import VueMarkdown from "vue-markdown-render";
 
 const authoritiesHelper = require("../helpers/libraryAuthorities.js");
 
@@ -93,7 +93,7 @@ export default {
   beforeMount() {
     this.getServices();
   },
-  components: { "layout-header": Header, VueMarkdownPlus }
+  components: { "layout-header": Header, VueMarkdown }
 };
 </script>
 
