@@ -1,9 +1,12 @@
 <template>
   <div>
-    <layout-header title="Stock summary" subtitle="Snapshot counts of library stock by item type" />
+    <layout-header
+      title="Stock summary"
+      subtitle="Snapshot counts of library stock by item type"
+    />
     <section>
       <v-container>
-        <vue-markdown-plus :source="mdText"></vue-markdown-plus>
+        <vue-markdown :source="mdText"></vue-markdown>
         <p class="text-center">
           <v-btn
             depressed
@@ -22,7 +25,7 @@
 <script>
 import Header from "../components/layout-header";
 import MarkDownData from "../markdown/stocksummarydata.md";
-import VueMarkdownPlus from "vue-markdown-plus";
+import VueMarkdown from "vue-markdown-render";
 
 export default {
   data() {
@@ -33,7 +36,7 @@ export default {
   methods: {},
   components: {
     "layout-header": Header,
-    VueMarkdownPlus
+    VueMarkdown
   }
 };
 </script>
